# assets/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Megazord API",
    "description": "OpenAPI for the Megazord platform",
    "contact": {
      "name": "GitHub repo",
      "url": "https://github.com/udaltsovk/megazord"
    },
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/",
      "description": "Local server"
    }
  ],
  "paths": {
...
    "/profiles/me": {
      "get": {
        "tags": [
          "Profiles"
        ],
        "description": "\n\n\n\n",
        "operationId": "get_current_profile",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Profile"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "participant": []
          },
          {
            "mentor": []
          },
          {
            "organizator": []
          }
        ]
      },
      "put": {
        "tags": [
          "Profiles"
        ],
        "description": "\n\n\n\n",
        "operationId": "upsert_current_profile",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertProfile"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Profile"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Profiles"
        ],
        "description": "\n\n\n\n",
        "operationId": "delete_current_profile",
        "responses": {
          "204": {
            "description": ""
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "participant": []
          },
          {
            "mentor": []
          },
          {
            "organizator": []
          }
        ]
      }
    },
    "/profiles/{profile_id}": {
      "get": {
        "tags": [
          "Profiles"
        ],
        "description": "\n\n\n\n",
        "operationId": "get_profile_by_id",
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string",
              "format": "ulid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Profile"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "participant": []
          },
          {
            "mentor": []
          },
          {
            "organizator": []
          }
        ]
      },
      "put": {
        "tags": [
          "Profiles"
        ],
        "description": "\n\n\n\n",
        "operationId": "upsert_profile_by_id",
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string",
              "format": "ulid"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertProfile"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Profile"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "organizator": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Profiles"
        ],
        "description": "\n\n\n\n",
        "operationId": "delete_profile_by_id",
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string",
              "format": "ulid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "organizator": []
          }
        ]
      }
    },
    "/profiles/me/image": {
      "get": {
        "tags": [
          "Profiles"
        ],
        "description": "\n\n\n\n",
        "operationId": "get_current_profile_image",
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "participant": []
          },
          {
            "mentor": []
          },
          {
            "organizator": []
          }
        ]
      },
      "put": {
        "tags": [
          "Profiles"
        ],
        "description": "\n\n\n\n",
        "operationId": "upsert_current_profile_image",
        "requestBody": {
          "description": "",
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UploadForm"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "413": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "415": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Profiles"
        ],
        "description": "\n\n\n\n",
        "operationId": "delete_current_profile_image",
        "responses": {
          "204": {
            "description": ""
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "participant": []
          },
          {
            "mentor": []
          },
          {
            "organizator": []
          }
        ]
      }
    },
    "/profiles/{profile_id}/image": {
      "get": {
        "tags": [
          "Profiles"
        ],
        "description": "\n\n\n\n",
        "operationId": "get_profile_image_by_id",
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string",
              "format": "ulid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "participant": []
          },
          {
            "mentor": []
          },
          {
            "organizator": []
          }
        ]
      },
      "put": {
        "tags": [
          "Profiles"
        ],
        "description": "\n\n\n\n",
        "operationId": "upsert_profile_image_by_id",
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string",
              "format": "ulid"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UploadForm"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "413": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "415": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "organizator": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Profiles"
        ],
        "description": "\n\n\n\n",
        "operationId": "delete_profile_image_by_id",
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string",
              "format": "ulid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "organizator": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "ApiError": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "Authentication"
            ],
            "properties": {
              "Authentication": {
                "type": "string",
                "enum": [
                  "NoAuthorizationHeader",
                  "InvalidCredentials",
                  "InvalidAuthMethod",
                  "MissingPermissions"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "Unauthorized"
            ],
            "properties": {
              "Unauthorized": {
                "type": "string"
              }
            }
          },
          {
            "type": "string",
            "enum": [
              "Forbidden"
            ]
          },
          {
            "type": "object",
            "required": [
              "ForbiddenWithMsg"
            ],
            "properties": {
              "ForbiddenWithMsg": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "NotFound"
            ],
            "properties": {
              "NotFound": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "AlreadyExists"
            ],
            "properties": {
              "AlreadyExists": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "PayloadTooLarge"
            ],
            "properties": {
              "PayloadTooLarge": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "UnsupportedMediaType"
            ],
            "properties": {
              "UnsupportedMediaType": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "Internal"
            ],
            "properties": {
              "Internal": {
                "type": "string"
              }
            }
          }
        ],
        "description": ""
      },
...
      "Profile": {
        "type": "object",
        "description": "",
        "required": [
          "id",
          "name",
          "surname",
          "city",
          "bio",
          "portfolio_urls"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "ulid",
            "description": ""
          },
          "name": {
            "type": "string",
            "description": "",
            "maxLength": 128,
            "minLength": 1
          },
          "surname": {
            "type": "string",
            "description": "",
            "maxLength": 128,
            "minLength": 1
          },
          "city": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1
          },
          "bio": {
            "type": "string",
            "description": "",
            "maxLength": 4096,
            "minLength": 0
          },
          "portfolio_urls": {
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 0
            },
            "description": ""
          }
        }
      },
      "UploadForm": {
        "type": "object",
        "required": [
          "file"
        ],
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "description": "Изображение, которое будет загружено.\nРазмер не должен превышать 5.7 МБ.\nРазрешённые MIME типы: `image/jpeg`, `image/pjpeg`, `image/png`, `image/webp`",
            "contentMediaType": "application/octet-stream"
          }
        }
      },
      "UpsertProfile": {
        "type": "object",
        "description": "",
        "required": [
          "name",
          "surname",
          "city",
          "bio",
          "portfolio_urls"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "",
            "maxLength": 128,
            "minLength": 1
          },
          "surname": {
            "type": "string",
            "description": "",
            "maxLength": 128,
            "minLength": 1
          },
          "city": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1
          },
          "bio": {
            "type": "string",
            "description": "",
            "maxLength": 4096,
            "minLength": 0
          },
          "portfolio_urls": {
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 0
            },
            "description": ""
          }
        }
      },
...
    },
    "securitySchemes": {
      "mentor": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "organizator": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "participant": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}